Moves components and css modules into separate folders, starts modal styling
