+++ client/components/customer_images/stylesCustomerImages.css
@@ -4,7 +4,7 @@
 }
 
 .customerImage {
-  composes: icon from './stylesComponents.css';
+  composes: icon from '../styled_components/stylesComponents.css';
   background-position: center;
   background-size: cover;
 }
