+++ client/components/modal/stylesModal.css
@@ -1,0 +1,61 @@
+.toGalleryWrapper {
+  composes: iconWrapper from '../styled_components/stylesComponents.css';
+  color: black;
+  font-size: 13px;
+  font-weight: bold;
+  text-decoration: none;
+  justify-content: flex-start;
+  margin: 15px 0 0 15px;
+
+}
+.toGallery {
+  composes: icon from '../styled_components/stylesComponents.css';
+  background-size: 500px;
+  background-position: -200px -4px;
+  width: 19px;
+  height: 19px;
+  display: inline-block;
+  margin-right: 5px;
+}
+
+.reviewWrapper {
+  width: 100%;
+  height: auto;
+  padding: 15px;
+  box-sizing: border-box;
+  flex-grow: 1;
+  display: flex;
+  align-items: stretch;
+}
+.review {
+  color: hotpink;
+}
+
+.largeImage {
+  background-color: #000;
+  background-position: center center;
+  background-repeat: no-repeat;
+  background-size: 100%;
+  flex: 1 1 auto;
+  display: flex;
+  width: 100%;
+  max-width: 400px;
+  align-items: stretch;
+  margin-right: 15px;
+}
+
+.arrowWrapper {
+  display: flex;
+  opacity: 0;
+}
+.arrowWrapper:hover {
+  opacity: 0.7;
+}
+.arrow {
+  composes: icon from '../styled_components/stylesComponents.css';
+  background-size: 800px;
+  background-position-y: -1445px;
+  height: 62px;
+  width: 40px;    
+  margin: 0 10px;
+}
